updated keepalive local variable from uint32_t to unsigned long (problem found in 64-bit machine)

+++ private/layer.h
@@ -1155,9 +1155,9 @@ error:
 
             if (status != idigi_callback_abort)
             {
-                uint32_t rx_keepalive = 0;
-                uint32_t tx_keepalive = 0;
-                uint32_t current_system_time;
+                unsigned long rx_keepalive = 0;
+                unsigned long tx_keepalive = 0;
+                unsigned long current_system_time;
                 /* check rx_keepalive and tx_keepalive timing */
                 status =  get_keepalive_timeout(idigi_ptr, &rx_keepalive, &tx_keepalive, &current_system_time);
                 if (rx_keepalive == 0 || tx_keepalive == 0 || status != idigi_callback_continue)
